commit version 0.5 spring-crud

+++ src/main/java/com/artyom/crud/entity/User.java
@@ -18,7 +18,7 @@ public class User {
     private Byte age;
 
     public static User newOf(UserRequest request) {
-        return new User(null, request.getName(), request.getLastName(), request.getAge());
+        return new User(null, request.getName(), request.getLastName(), request.getAge().byteValue());
     }
 
     public User() {
